Admin Layouts, authentication fix, feedback messages names fix

+++ app/controllers/concerns/authentication.rb
@@ -37,7 +37,11 @@ module Authentication
     end
 
     def after_authentication_url
-      session.delete(:return_to_after_authenticating) || root_url
+      if Current.user&.admin?
+        admin_root_path
+      else
+        session.delete(:return_to_after_authenticating) || root_url
+      end
     end
 
     def start_new_session_for(user)
